storage: rename controller's healthcheck to collection_status
