+++ src/storage-controller/src/collection_status.rs
